+++ spec/snippetPreviewSpec.js
@@ -1,0 +1,24 @@
+var SnippetPreview = require("../js/snippetPreview.js");
+require("../js/app.js");
+
+describe("The snippet preview constructor", function() {
+	it("accepts an App object as an opts property", function() {
+		var mockApp = {
+			rawData: {
+				snippetTitle: "",
+				snippetCite: "",
+				snippetMeta: ""
+			}
+		};
+		// Makes lodash think this is a valid HTML element
+		var mockElement = [];
+		mockElement.nodeType = 1;
+
+		var snippetPreview = new SnippetPreview({
+			analyzerApp: mockApp,
+			targetElement: mockElement
+		});
+
+		expect(snippetPreview.refObj).toBe(mockApp);
+	})
+});